requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
anywidget==0.9.21
asttokens==3.0.1
bqplot==0.12.45
branca==0.8.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
earthengine-api==1.7.4
eerepr==0.1.2
executing==2.2.1
fastapi==0.128.0
folium==0.20.0
fonttools==4.61.1
future==1.0.0
geemap==0.36.6
geocoder==1.38.1
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-storage==3.8.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
h11==0.16.0
httplib2==0.31.1
idna==3.11
ipyevents==2.0.4
ipyfilechooser==0.6.0
ipyleaflet==0.20.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
jedi==0.19.2
Jinja2==3.1.6
jupyter-leaflet==0.20.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
narwhals==2.15.0
numpy==2.4.1
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
plotly==6.5.2
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==6.33.4
psygnal==0.15.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.1
pyperclip==1.11.0
pyproj==3.7.2
pyshp==3.0.3
python-box==7.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
ratelim==0.1.6
requests==2.32.5
rsa==4.9.1
scooby==0.11.0
shapely==2.1.2
six==1.17.0
stack-data==0.6.3
starlette==0.50.0
traitlets==5.14.3
traittypes==0.2.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
wcwidth==0.2.14
widgetsnbextension==4.0.15
xyzservices==2025.11.0
